feature(Least Significant Bit): Create header containing data length
The header is placed at the beginning of the image, and it's size is calculated depending on the image dimensions, this way the header is always found. The hidden data is then accompanied by a flag bit, "1" if data is a file and "0" if it was a simple message. binary_to_file function was added to complete the decoding function. The data is now also checked to see if it can fit in the cover image. This commit has yet to be tested.